# Conflicts:
#	src/Assetic/Filter/Sass/SassFilter.php
#	src/Assetic/Filter/TypeScriptFilter.php

+++ tests/Assetic/Test/Cache/FilesystemCacheTest.php
@@ -32,6 +32,8 @@ class FilesystemCacheTest extends \PHPUnit_Framework_TestCase
 
         $cache->remove('foo');
         $this->assertFalse($cache->has('foo'));
+
+        $this->removeDir($dir);
     }
 
     public function testSetCreatesDir()
@@ -45,11 +47,15 @@ class FilesystemCacheTest extends \PHPUnit_Framework_TestCase
         $this->assertFileExists($dir.'/foo');
 
         $this->removeDir($dir);
+        rmdir(sys_get_temp_dir().'/assetic');
     }
 
     private function removeDir($dir)
     {
         array_map('unlink', glob($dir.'/*'));
-        @rmdir($dir);
+
+        if (is_dir($dir)) {
+            rmdir($dir);
+        }
     }
 }
